Base functionality for the pages
This is super basic but the app is now "working" in the sense that the
pages show the the data from the json file and we can swipe new
questions. More is to be done such as randomize the questions we get,
get buttons to work as intended, visuals, evaluate state management and
more.