Updating text widget with a button press
This is a rough implementation but for now, when a button is a pressed
we set the state of the text widget with the value of the _difficultly
value. This will allow us to pull different sets of questions from the
json file. In the future I want another json file for the secondary page
and to make this more robust because this is prone for failure. But we
are barely in alpha now so this will be fine.