Now supporting new plugin API's
Updated MainActivity files to use FlutterActivity. The contents in the main function were not being used so they are removed to deal with this migration.
Updated manifest file to use embedding v2 and new styles which will be updated on another branch, this is more of a placeholder for now. I was following instructions here: https://github.com/flutter/flutter/wiki/Upgrading-pre-1.12-Android-projects